+++ tests/exclusions/expected/Module/java/Tags.java
@@ -1,0 +1,15 @@
+package com.example.module.model;
+
+import lombok.*;
+
+// An example for tags etc.
+@Builder
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+public class Tags {
+  private String id;
+}
+
+