+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Added x-reference-type property to define the type of the reference
 - Static properties are ignored in java validation
+- Added exclusion on application level
 
 ### Changed